requirements.txt
coc.py==2.0.0
discord.py==1.7.3
discord-py-interactions==3.0.2
Flask==2.0.2
Pillow==8.4.0
requests==2.26.0
topggpy==1.4.0
